Allow non-integral X-Ray sampling rates (#803)

+++ config.go
@@ -125,5 +125,5 @@ type Config struct {
 	} `yaml:"veneur_metrics_scopes"`
 	XrayAddress          string   `yaml:"xray_address"`
 	XrayAnnotationTags   []string `yaml:"xray_annotation_tags"`
-	XraySamplePercentage int      `yaml:"xray_sample_percentage"`
+	XraySamplePercentage float64  `yaml:"xray_sample_percentage"`
 }
